Avoid modifying the JavaScriptEncoder on SystemTextJsonOutputFormatter
By default, STJ encodes most non-ASCII characters which is different from Newtonsoft.Json's
defaults. When we first defaulted to STJ in 3.1, MVC attempted to minimize this difference by
using a more compatible (unsafe-relaxed) encoding scheme if a user hadn't explicitly configured one via
JsonOptions.

As noted in https://github.com/dotnet/aspnetcore/issues/38720, this causes issues if a
JsonSerializerContext is configured. This PR changes the output formatter to no longer
change the JavaScriptEncoder. Users can manually configure the unsafe-relaxed encoding
globally if they understand the consequences of doing so.

Contributes to https://github.com/dotnet/aspnetcore/issues/38720
