+++ src/Mvc/Mvc.NewtonsoftJson/test/NewtonsoftJsonOutputFormatterTest.cs
@@ -118,6 +118,37 @@ public class NewtonsoftJsonOutputFormatterTest : JsonOutputFormatterTestBase
     }
 
     [Fact]
+    public async Task WriteResponseBodyAsync_Encodes()
+    {
+        // Arrange
+        var formatter = GetOutputFormatter();
+        var expectedContent = "{\"key\":\"Hello \\n <b>Wörld</b>\"}";
+        var content = new { key = "Hello \n <b>Wörld</b>" };
+
+        var mediaType = MediaTypeHeaderValue.Parse("application/json; charset=utf-8");
+        var encoding = CreateOrGetSupportedEncoding(formatter, "utf-8", isDefaultEncoding: true);
+
+        var body = new MemoryStream();
+        var actionContext = GetActionContext(mediaType, body);
+
+        var outputFormatterContext = new OutputFormatterWriteContext(
+            actionContext.HttpContext,
+            new TestHttpResponseStreamWriterFactory().CreateWriter,
+            typeof(string),
+            content)
+        {
+            ContentType = new StringSegment(mediaType.ToString()),
+        };
+
+        // Act
+        await formatter.WriteResponseBodyAsync(outputFormatterContext, Encoding.GetEncoding("utf-8"));
+
+        // Assert
+        var actualContent = encoding.GetString(body.ToArray());
+        Assert.Equal(expectedContent, actualContent);
+    }
+
+    [Fact]
     public async Task ChangesTo_SerializerSettings_AffectSerialization()
     {
         // Arrange
@@ -547,6 +578,56 @@ public class NewtonsoftJsonOutputFormatterTest : JsonOutputFormatterTestBase
             cancellationToken.ThrowIfCancellationRequested();
             yield return 1;
         }
+    }
+
+    public static TheoryData<string, string, bool> WriteCorrectCharacterEncoding
+    {
+        get
+        {
+            var data = new TheoryData<string, string, bool>
+            {
+                { "This is a test 激光這兩個字是甚麼意思 string written using utf-8", "utf-8", true },
+                { "This is a test 激光這兩個字是甚麼意思 string written using utf-16", "utf-16", true },
+                { "This is a test 激光這兩個字是甚麼意思 string written using utf-32", "utf-32", false },
+                { "This is a test æøå string written using iso-8859-1", "iso-8859-1", false },
+            };
+
+            return data;
+        }
+    }
+
+    [Theory]
+    [MemberData(nameof(WriteCorrectCharacterEncoding))]
+    public async Task WriteToStreamAsync_UsesCorrectCharacterEncoding(
+       string content,
+       string encodingAsString,
+       bool isDefaultEncoding)
+    {
+        // Arrange
+        var formatter = GetOutputFormatter();
+        var expectedContent = "\"" + content + "\"";
+        var mediaType = MediaTypeHeaderValue.Parse($"application/json; charset={encodingAsString}");
+        var encoding = CreateOrGetSupportedEncoding(formatter, encodingAsString, isDefaultEncoding);
+
+        var body = new MemoryStream();
+        var actionContext = GetActionContext(mediaType, body);
+
+        var outputFormatterContext = new OutputFormatterWriteContext(
+            actionContext.HttpContext,
+            new TestHttpResponseStreamWriterFactory().CreateWriter,
+            typeof(string),
+            content)
+        {
+            ContentType = new StringSegment(mediaType.ToString()),
+        };
+
+        // Act
+        await formatter.WriteResponseBodyAsync(outputFormatterContext, Encoding.GetEncoding(encodingAsString));
+
+        // Assert
+        var actualContent = encoding.GetString(body.ToArray());
+        Assert.Equal(expectedContent, actualContent, StringComparer.OrdinalIgnoreCase);
+        Assert.True(body.CanWrite, "Response body should not be disposed.");
     }
 
     private class TestableJsonOutputFormatter : NewtonsoftJsonOutputFormatter
