+++ src/Mvc/test/Mvc.FunctionalTests/JsonOutputFormatterTestBase.cs
@@ -1,4 +1,4 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
+// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System.Net;
@@ -105,7 +105,7 @@ public abstract class JsonOutputFormatterTestBase<TStartup> : IClassFixture<MvcT
 
         // Assert
         await response.AssertStatusCodeAsync(HttpStatusCode.OK);
-        Assert.Equal("\"Une bête de cirque\"", await response.Content.ReadAsStringAsync());
+        Assert.Equal("\"Une b\\u00EAte de cirque\"", await response.Content.ReadAsStringAsync());
     }
 
     [Fact]
